Completed Models for JSON API Request & Responses; Completed basic HomeScene with no transaction history yet;

+++ mdt/Services/ResponseModels/ReadBalanceResponse.swift
@@ -10,7 +10,7 @@ import Foundation
 struct ReadBalanceResponse: Codable {
     let status: String
     let accountNo: String?
-    let balance: Int?
+    let balance: Double?
     let error: String?
     
     enum CodingKeys: String, CodingKey {
