+++ mdt/Services/ResponseModels/ReadTransactionsResponse.swift
@@ -33,7 +33,7 @@ extension ReadTransactionsResponse {
 
 struct Transaction: Codable {
     let transactionId: String
-    let amount: Int
+    let amount: Double
     let transactionDate: String
     let description: String
     let transactionType: String
